server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
rembg==2.0.50
sqlmodel==0.0.14
sqlite-utils==3.35
opencv-python-headless==4.8.1.78
numpy==1.24.3

# MobileSAM dependencies
onnxruntime==1.16.3
torch==2.1.1
torchvision==0.16.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
